requirements.txt
flask
flask-restful
flask-sqlalchemy
psycopg2-binary
flask-cors
flask-bcrypt
flask-migrate
flask-limiter
flask-jwt-extended
flasgger
requests
python-dotenv
openai
instagrapi
redis
pandas
gunicorn
celery
sentry-sdk[flask]
